api/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
web3==5.24.0
PyMySQL==1.0.2
cryptography==36.0.1
flask_migrate==3.1.0
python-decouple==3.5
